ppc64_cpu: test_attr return 2 when file missing
The old logic was to assume the attribute is readable when the sysfs
file does not exist. This makes it hard to detect when subcores are
available.

However, the existing code relies on not returning an error when
iterating through the CPUs in sysfs. Compromise by returning 2, which
doesn't break existing code, and allows new users to ask precisely what
the state of the attribute is.

Signed-off-by: Joel Stanley <joel@jms.id.au>