ppc64_cpu: work around bad sysfs online behaviour
The sysfs file /sys/devices/system/cpu/cpuXX/online will return EINVAL
if set to 1 when already online, or 0 when already offline. This is not
a fatal error: it is normal behaviour when setting the smt mode.

As we are stuck with the sysfs interface, work around it by looking at
errno after setting the CPU state, and ignoring the failure if errno is
set to EINVAL.

While we are here, merge the two for loops into one to avoid duplicating
the error checking code.

Signed-off-by: Joel Stanley <joel@jms.id.au>