ppc64_cpu: handle offline cpus when setting attributes
When a CPU is offline, sysfs files are removed from the
/sys/devices/system/cpu/cpuXX directory. The dscr and smt_snooze_delay
settings are set on every CPU, but for offline CPUs the sysfs file will
not exist. Detect when set_attribute has errored out due to a missing
file, and ignore it.

The only callers of set_system_attribute are dscr and smt_snooze_delay,
so special casing for their expected behaviour is okay.

Signed-off-by: Joel Stanley <joel@jms.id.au>