ppc64_cpu: Remove smt_snooze_delay
The kernel has not supported setting the snooze delay through this API
since at least 2014 (v3.14). The sysfs file is present for
compatibility, but it has no effect.

This removes the ability for ppc64_cpu to detect the smt state out of
of the box as there is no obvious indication provided to userspace.
However, this is not required by the --smt flag, so we can remove that
code without loss of functionality.

Signed-off-by: Joel Stanley <joel@jms.id.au>